Initial refactoring of code in line with SOLID principles
