+++ Java/TelemetrySystem/pom.xml
@@ -13,9 +13,17 @@
 	<artifactId>TelemetrySystem</artifactId>
 
 	<dependencies>
+
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-engine</artifactId>
+			<artifactId>junit-jupiter-api</artifactId>
+			<version>5.8.2</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>4.5.1</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
