+++ Java/TelemetrySystem/src/main/java/tddmicroexercises/telemetrysystem/TelemetryClient/ITelemetryClient.java
@@ -1,0 +1,7 @@
+package tddmicroexercises.telemetrysystem.TelemetryClient;
+
+public interface ITelemetryClient {
+
+    void send(String message);
+    String receive();
+}
