+++ Java/TelemetrySystem/src/main/java/tddmicroexercises/telemetrysystem/TelemetryConnection/Connection.java
@@ -1,0 +1,9 @@
+package tddmicroexercises.telemetrysystem.TelemetryConnection;
+
+public interface Connection {
+    boolean getConnectionStatus();
+
+    void connect(String connectionString);
+
+    void disconnect();
+}
