+++ Java/TelemetrySystem/src/main/java/tddmicroexercises/telemetrysystem/TelemetryDiagnostics/ITelemetryDiagnosticControls.java
@@ -1,0 +1,7 @@
+package tddmicroexercises.telemetrysystem.TelemetryDiagnostics;
+
+public interface ITelemetryDiagnosticControls {
+    String getDiagnosticInfo();
+    void setDiagnosticInfo(String diagnosticInfo);
+    void checkTransmission() throws Exception;
+}
