+++ Java/TelemetrySystem/src/test/java/tddmicroexercises/TelemetryDiagnostics/TelemetryDiagnosticControlsTest.java
@@ -1,0 +1,13 @@
+package tddmicroexercises.TelemetryDiagnostics;
+
+import org.junit.jupiter.api.Test;
+
+public class TelemetryDiagnosticControlsTest
+{
+
+    @Test
+    public void CheckTransmission_should_send_a_diagnostic_message_and_receive_a_status_message_response()
+    {
+    }
+
+}
