--- Java/TelemetrySystem/src/test/java/tddmicroexercises/telemetrysystem/TelemetryDiagnosticControlsTest.java
@@ -1,13 +1,0 @@
-package tddmicroexercises.telemetrysystem;
-
-import org.junit.jupiter.api.Test;
-
-public class TelemetryDiagnosticControlsTest
-{
-    
-	@Test
-    public void CheckTransmission_should_send_a_diagnostic_message_and_receive_a_status_message_response()
-    {
-    }
-
-}
